[FIX] Use signed integer types for CLI parameters - change args.min_var_length from uint32_t to int32_t - change args.max_var_length from uint32_t to int32_t - change args.max_tol_inserted_length from uint32_t to int32_t - change args.max_overlap from uint32_t to int32_t - expand CLI documentation (noting that params need non-negative value) - expand documentation of all functions accepting these arguments  (noting that it expects a non-negative value)  - explicitly check for non-negative value and terminate if negative  - add test for this check with negative parameter  - replace abs with std::abs in analyze_sa_tag_method.cpp
Reason:
When comparing signed with unsigned values
(e.g. int32 and uint32), C++ implicitly converts the signed
type into an unsigned type which leads to unexpected
behavior (e.g. -1 gets converted to a large positive value).

As a solution, we should not use unsigned types for quantities
on which we might want to perform mathematical operations
or comparisons. To make sure that the parameters do not take
negative values we add an explicit check that terminates the
program if the user gave a negative value.

+++ test/api/detection_test.cpp
@@ -27,7 +27,7 @@ TEST(junction_detection, cigar_string_simple_del)
     // Deletion smaller than minimum variant size
     {
         std::vector<Junction> junctions_res{};
-        uint64_t const min_var_length = 10;
+        int32_t const min_var_length = 10;
         analyze_cigar(read_name, chromosome, query_start_pos, cigar_string, seq, junctions_res, min_var_length);
 
         EXPECT_EQ(junctions_res.size(), 0);
@@ -36,7 +36,7 @@ TEST(junction_detection, cigar_string_simple_del)
     // Deletion larger than minimum variant size
     {
         std::vector<Junction> junctions_res{};
-        uint64_t const min_var_length = 5;
+        int32_t const min_var_length = 5;
         analyze_cigar(read_name, chromosome, query_start_pos, cigar_string, seq, junctions_res, min_var_length);
 
         Breakend new_breakend_1 {chromosome, 15, strand::forward};
@@ -66,7 +66,7 @@ TEST(junction_detection, cigar_string_del_padding)
     seqan3::dna5_vector seq = {"GGGCTCATCGATCGATTTCGGATCGGGGGGCCCCCATTTTAAACGGCCCC"_dna5};
 
     std::vector<Junction> junctions_res{};
-    uint64_t const min_var_length = 5;
+    int32_t const min_var_length = 5;
     analyze_cigar(read_name, chromosome, query_start_pos, cigar_string, seq, junctions_res, min_var_length);
 
     Breakend new_breakend_1 {chromosome, 15, strand::forward};
@@ -94,7 +94,7 @@ TEST(junction_detection, cigar_string_simple_ins)
     seqan3::dna5_vector seq = {"GGGCTCATCGATCGATTTCGGATCGGGGGGCCCCCATTTTAAACGGCCCC"_dna5};
 
     std::vector<Junction> junctions_res{};
-    uint64_t const min_var_length = 5;
+    int32_t const min_var_length = 5;
     analyze_cigar(read_name, chromosome, query_start_pos, cigar_string, seq, junctions_res, min_var_length);
 
     Breakend new_breakend_1 {chromosome, 9, strand::forward};
@@ -122,7 +122,7 @@ TEST(junction_detection, cigar_string_ins_hardclip)
     seqan3::dna5_vector seq = {"GGGCTCATCGATCGATTTCGGATCGGGGGGCCCCCATTTTAAACGGCCCC"_dna5};
 
     std::vector<Junction> junctions_res{};
-    uint64_t const min_var_length = 5;
+    int32_t const min_var_length = 5;
     analyze_cigar(read_name, chromosome, query_start_pos, cigar_string, seq, junctions_res, min_var_length);
 
     Breakend new_breakend_1 {chromosome, 9, strand::forward};
